fix all comments
