+++ app/javascript/stylesheets/feedback_form.scss
@@ -1,7 +1,3 @@
-abbr[title="required"] {
-  display: none;
-}
-
 .feedback-form {
   .form-check {
     display: flex !important;
@@ -9,5 +5,5 @@ abbr[title="required"] {
 }
 
 .form-check-label {
-    margin-left: 1em;
+  margin-left: 1em;
 }
