+++ app/javascript/stylesheets/style.scss
@@ -1,3 +1,7 @@
+abbr[title="required"] {
+  display: none;
+}
+
 .home-page {
   padding-left: 2rem;
   padding-right: 5rem;
