+++ app/views/feedbacks/_form.html.slim
@@ -1,14 +1,14 @@
 section.team-section.col
-  = simple_form_for @feedback, url:'contact_us', method: :post do |f|
+  = simple_form_for @feedback, url: feedback_path, method: :post do |f|
     .container.my-5  
     .card.m-2
       .card-header[style="background-color:#734dbb"]
         p.h5.text-center.text-light.font-weight-bold.mb-0
           | Feedback
-
+       
           .field.mb-3
             = f.input :name, label: "Your name:", class: 'form-control'
-
+          
           .field.mb-3
             = f.input :email, label: "Your Email:", placeholder: 'example@email.com'
 
@@ -21,6 +21,9 @@ section.team-section.col
 
           .field.mb-3
             = f.input :message, label: "Reason: ", placeholder: "Write message ... ", as: :text, :html_input => {:cols => 5, :rows => 15}
+          
+          .field  
+            = hidden_field_tag :authenticity_token, form_authenticity_token
 
           .action
             = f.submit "Send"
