+++ app/views/layouts/application.html.slim
@@ -11,7 +11,7 @@ html
     = javascript_pack_tag 'application', 'data-turbolinks-track': 'reload'
 
   body
-    = render "header"
+    = render "static_pages/header"
 
     main
       = render 'shared/flashes'
