+++ app/views/static_pages/_header.html.slim
@@ -3,9 +3,9 @@
   /! Nav tabs
   nav.nav.justify-content-center
     = link_to 'Home', root_path, class: "nav-link text_color active", aria_current: "page"
-    a.nav-link.text_color Pricing
-    a.nav-link.text_color href="#"  About
-    a.nav-link.text_color href="contact_us"  Contact Us
+    = link_to 'Pricing', '#', class: "nav-link text_color"
+    = link_to 'About', '#', class: "nav-link text_color"
+    = link_to 'Contact Us', feedback_path, class: "nav-link text_color"
 
 .d-grid.gap-2.d-md-flex.justify-content-md-end.menu-btn
   button.btn.btn-primary.me-md-2.btn-color Log in
